backend/requirements.txt
Django>=4.2
djangorestframework
django-cors-headers
mysqlclient
boto3
PyJWT>=2.4.0